feat: Enable auto-wiring
tl;dr: Removes the need for registering factory functions for
dependencies that only resolve to classes.

Refactored the resolver to resolve the full dependency tree of a
dependency in one go, rather than on each dependency access. This still
happens at dependency-lookup, but internally it evaluates each
sub-dependency at once.

This allows us to auto-wire dependencies at request-time, by creating a
factory function for the dependency and sub-dependencies, if not already
registered in the container.

Started to add some detection for non-class dependencies that cannot be
populated automatically, but I think there's a few more cases like this
still present.

Non-class dependencies with default values, aren't yet populated. e.g.:
    __init__(self, foo: str = "default")